Add addon_config:rw to config.yaml
This mounts a /config directory inside the container. I've turned on rw so that we can write log files to this directory. The directory is mounted from /addon_configs/<repo_slug>/ so we can check the files from inside home assistant at that path.

See https://developers.home-assistant.io/docs/add-ons/configuration/#add-on-advanced-options