Expose log files from container
This mounts a /config directory inside the container. I've turned on rw so that we can write log files to this directory. The directory is mounted from /addon_configs/<repo_slug>/ so we can check the files from inside home assistant at that path.

The icecast config has been updated to write it's log files to this /config directory.

See https://developers.home-assistant.io/docs/add-ons/configuration/#add-on-advanced-options